Correct words are displaying
